TASK_1/requirements.txt
pymongo[srv]==3.6
bson==0.5.10
python-dotenv